+++ src/components/Card/Card.vue
@@ -1,5 +1,5 @@
 <template>
-  <div class="card" :class="miter ? 'miter' : ''">
+  <div class="card" :class="{ ['shadow-' + shadow]: shadow, miter: miter }">
     <div>
       <header v-if="$slots.header">
         <slot name="header" />
@@ -30,8 +30,8 @@ import { computed } from 'vue'
 class Props {
   height = prop<string>({ default: null })
   width = prop<string>({ default: null })
-  shadow = prop<boolean>({ default: false })
-  miter = prop<boolean>({ default: false })
+  shadow = prop<string>({ default: null })
+  miter = prop<boolean>({ default: false, type: Boolean })
 }
 
 // This alias is needed because vue-class-component isn't yet compatible with storybook